Split project up into multiple packages
