+++ packages/jest-snapshots-svg/_tests/__fixtures__/artwork-grid.json
