+++ packages/jest-snapshots-svg/_tests/__fixtures__/welcome.json
