+++ packages/jest-snapshots-svg/_tests/_component-to-node.test.ts
