+++ packages/jest-snapshots-svg/_tests/_node-instance-count.test.tsx
