+++ packages/jest-snapshots-svg/_tests/_node-to-svg.test.ts
