+++ packages/jest-snapshots-svg/_tests/example_layouts/_align-items.test.tsx
