+++ packages/jest-snapshots-svg/_tests/example_layouts/_flex-direction.test.tsx
