+++ packages/jest-snapshots-svg/_tests/example_layouts/_justify-contents.test.tsx
