+++ packages/jest-snapshots-svg/_tests/example_layouts/_position.test.tsx
