+++ packages/jest-snapshots-svg/_tests/example_layouts/_text.test.tsx
