+++ packages/jest-snapshots-svg/_tests/render.test.tsx
