+++ packages/jest-snapshots-svg/ambient.ts
@@ -1,0 +1,8 @@
+// Add our matcher to Jest's definitions
+
+declare namespace jest {
+  interface Matchers<R> {
+    /** Checks and sets up SVG rendering for React Components. */
+    toMatchSVGSnapshot(width: number, height: number): void;
+  }
+}
