+++ packages/jest-snapshots-svg/external-api.d.ts
