+++ packages/jest-snapshots-svg/package.json
@@ -1,0 +1,65 @@
+{
+  "name": "jest-snapshots-svg",
+  "version": "0.0.21",
+  "description": "Generate SVG Snapshots of React Native Component Trees",
+  "main": "build/index.js",
+  "files": [
+    "build/*"
+  ],
+  "typings": "build/external-api.d.ts",
+  "author": "Orta Therox <orta.therox@gmail.com> & Art.sy Inc",
+  "license": "MIT",
+  "devDependencies": {
+    "@types/jest": "^20.0.8",
+    "@types/node": "^7.0.31",
+    "babel-cli": "^6.24.1",
+    "babel-jest": "^20.0.3",
+    "danger": "^0.18.0",
+    "husky": "^0.13.3",
+    "jest": "^21.0.1",
+    "lint-staged": "^3.2.5",
+    "prop-types": "^15.5.10",
+    "react": "16.0.0-alpha.6",
+    "react-native": "^0.44.0",
+    "react-test-renderer": "^15.5.4",
+    "ts-jest": "^21.0.0",
+    "ts-node": "^3.0.0",
+    "tslint": "^5.2.0",
+    "typescript": "^2.3.2"
+  },
+  "scripts": {
+    "type-check": "tsc --noEmit",
+    "build": "tsc && cp src/external-api.d.ts build",
+    "watch": "tsc --watch",
+    "lint": "tslint 'src/**/*.{ts,tsx}'",
+    "precommit": "lint-staged",
+    "prepush": "yarn build && yarn jest"
+  },
+  "lint-staged": {
+    "*.@(ts|tsx)": [
+      "tslint --fix",
+      "git add"
+    ]
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js"
+    ],
+    "transform": {
+      "^.+\\.js$": "<rootDir>/node_modules/babel-jest",
+      ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+    },
+    "testRegex": "(.test)\\.(ts|tsx)$",
+    "testPathIgnorePatterns": [
+      "\\.snap$",
+      "<rootDir>/node_modules/"
+    ],
+    "cacheDirectory": ".jest/cache",
+    "preset": "react-native"
+  },
+  "dependencies": {
+    "project-name-svg": "0.0.21"
+  }
+}
