+++ packages/project-name-svg/index.ts
@@ -1,0 +1,7 @@
+import layoutNode, { Component } from 'project-name-core'
+import treeToSVG from "./tree-to-svg"
+
+export default (root: Component, width: number, height: number): string => {
+    const settings = { width, height }
+    return treeToSVG(layoutNode(root, settings), settings)
+}
