+++ packages/project-name-svg/whitespace.ts
